Refactor updates
