+++ minihass/device.py
@@ -18,11 +18,10 @@ class Device:
     :class:`Device` object, although it it technically possible to create multiple
     objects in one instance of CircuitPython using this module.
 
-    .. caution:: The MQTT client used to create a :class:`Device` object must not be
-        connected at the time of instantiation. The devices uses a `Last Will and
-        Testament`_ message to mark the device and its entities as ``unavailable`` after
-        losing its connection to the MQTT broker, and the LWT is sent as part of the
-        connection process.
+    .. note:: The :class:`Device` object provides a `Last Will and Testament`_ message
+        to mark the device and its entities as ``unavailable`` after losing its
+        connection to the MQTT broker. This must be set as part of
+        :meth:`adafruit_minimqtt.adafruit_minimqtt.MQTT.connect()`.
 
     .. _Last Will and Testament: https://www.hivemq.com/blog/mqtt-essentials-part-9-last-will-and-testament/
